feat(reindex): refactor SecondaryIndexDocumentMissingError handling and update imports
